pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
pygments = "^2.17.2"
psycopg2-binary = "^2.9.9"
django-silk = "^5.1.0"
django-filter = "^24.2"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
django-elasticsearch-dsl = "^8.0"
celery = {extras = ["redis"], version = "^5.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
